Rename zbus_channels.def files to regular .h files
